requirements.txt
python-dotenv==1.0.1
dbt-core==1.8.1
dbt-postgres==1.8.1
GitPython==3.1.43